+++ build_runner/test/generate/build_test.dart
@@ -33,6 +33,46 @@ void main() {
       placeholderIdsFor(buildPackageGraph({rootPackage('a'): []}));
 
   group('build', () {
+    test('can log within a buildFactory', () async {
+      await testBuilders(
+        [
+          apply(
+              '',
+              [
+                (_) {
+                  log.info('I can log!');
+                  return new TestBuilder(
+                      buildExtensions: appendExtension('.1'));
+                }
+              ],
+              toRoot(),
+              isOptional: true,
+              hideOutput: false),
+        ],
+        {'a|web/a.txt': 'a'},
+      );
+    });
+
+    test('throws an error if the builderFactory fails', () async {
+      expect(
+          () async => await testBuilders(
+                [
+                  apply(
+                      '',
+                      [
+                        (_) {
+                          throw 'some error';
+                        }
+                      ],
+                      toRoot(),
+                      isOptional: true,
+                      hideOutput: false),
+                ],
+                {'a|web/a.txt': 'a'},
+              ),
+          throwsA('builderFactory did not return a builder.'));
+    });
+
     group('with root package inputs', () {
       test('one phase, one builder, one-to-one outputs', () async {
         await testBuilders(